StringDictionary.ro.xaml
I've created the Romanian translation as best as I could. Problem is some terms are equal in Romanian as they are in English so where you see there is no change (eg. Drive same as drive). 

I'll try to go over when I have some more free time to improve the translation. I am new to this translation GitHUB thing.